Reviewed by Mitz.
        Small tweak.

        * plugins/win/PluginStreamWin.cpp:
        (WebCore::PluginStreamWin::startStream):



Canonical link: https://commits.webkit.org/19878@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@25400 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebCore/plugins/win/PluginStreamWin.cpp
@@ -154,7 +154,7 @@ void PluginStreamWin::startStream()
     m_stream.headers = m_headers.data();
     m_stream.pdata = 0;
     m_stream.ndata = this;
-    m_stream.end = max(expectedContentLength, static_cast<long long>(0));
+    m_stream.end = max(expectedContentLength, 0LL);
     m_stream.lastmodified = m_resourceResponse.lastModifiedDate();
     m_stream.notifyData = m_notifyData;
